A brave new world.
Finally backed up the old versions of FullTransform and Shapelet,
still not confident 100% that the changes to the distance caching is correct.
More testing, which is why I've opted to keep the old verison which I've semi-merged

A few changes to classes which relied on the old version to ensure they still work.
A little more work to be done on getting everything upto date.

- TODO: work on threaded/caching
- TODO: work on re-implementing a few of the old methods for opCount, low prio!